Clag database creation against endpoint is working
Currently setting up a Priority queue for callable in order to make the service call the backend. Testing has been difficult but managed a bit. A bunch of stuff within clag that will enable automatic parsing & DB creation. There is a need to work on the Service and its lifecycle before moving forward.